Changed media configuration to work with Django 1.3 staticfiles
Media files were moved to static/inlinetrans/ and MEDIA_URL parameter was added
to settings.py. The URL path can be configured with INLINETRANS_MEDIA_URL parameter.
Also fixed file paths in inlinetrans_header.html template. Note that jQuery is not
bundled with inlinetrans, so inlinetrans should not add a script-tag for jQuery.